# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CarModPicker",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/cars/": {
      "post": {
        "tags": [
          "cars"
        ],
        "summary": "Create Car",
        "operationId": "create_car_api_cars__post",
        "parameters": [
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarRead"
                }
              }
            }
          },
          "400": {
            "description": "Car already exists"
          },
          "403": {
            "description": "Not authorized to create a car"
          },
          "402": {
            "description": "Subscription limit reached"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cars/{car_id}": {
      "get": {
        "tags": [
          "cars"
        ],
        "summary": "Read Car",
        "operationId": "read_car_api_cars__car_id__get",
        "parameters": [
          {
            "name": "car_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Car Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarRead"
                }
              }
            }
          },
          "404": {
            "description": "Car not found"
          },
          "403": {
            "description": "Not authorized to access this car"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "cars"
        ],
        "summary": "Update Car",
        "operationId": "update_car_api_cars__car_id__put",
        "parameters": [
          {
            "name": "car_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Car Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CarUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarRead"
                }
              }
            }
          },
          "404": {
            "description": "Car not found"
          },
          "403": {
            "description": "Not authorized to update this car"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "cars"
        ],
        "summary": "Delete Car",
        "operationId": "delete_car_api_cars__car_id__delete",
        "parameters": [
          {
            "name": "car_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Car Id"
            }
          },
          {
            "name": "access_token",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Access Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CarRead"
                }
              }
            }
          },
          "404": {
            "description": "Car not found"
          },
          "403": {
            "description": "Not authorized to delete this car"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/cars/user/{user_id}": {
      "get": {
        "tags": [
          "cars",
          "cars"
        ],
        "summary": "Read Cars By User",
        "description": "Retrieve all cars owned by a specific user with pagination.",
        "operationId": "read_cars_by_user_api_cars_user__user_id__get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of cars to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of cars to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of cars to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of cars to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CarRead"
                  },
                  "title": "Response Read Cars By User Api Cars User  User Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/build-lists/car/{car_id}": {
      "get": {
        "tags": [
          "build_lists",
          "build_lists"
        ],
        "summary": "Read Build Lists By Car",
        "description": "Retrieve all build lists associated with a specific car with pagination.",
        "operationId": "read_build_lists_by_car_api_build_lists_car__car_id__get",
        "parameters": [
          {
            "name": "car_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Car Id"
            }
          },
          {
            "name": "skip",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "description": "Number of build lists to skip",
              "default": 0,
              "title": "Skip"
            },
            "description": "Number of build lists to skip"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 1000,
              "minimum": 1,
              "description": "Maximum number of build lists to return",
              "default": 100,
              "title": "Limit"
            },
            "description": "Maximum number of build lists to return"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/BuildListRead"
                  },
                  "title": "Response Read Build Lists By Car Api Build Lists Car  Car Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuildListRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "car_id": {
            "type": "integer",
            "title": "Car Id"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "car_id"
        ],
        "title": "BuildListRead"
      },
...
      "CarCreate": {
        "properties": {
          "make": {
            "type": "string",
            "title": "Make"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "trim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trim"
          },
          "vin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vin"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "required": [
          "make",
          "model",
          "year"
        ],
        "title": "CarCreate"
      },
      "CarRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "make": {
            "type": "string",
            "title": "Make"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "year": {
            "type": "integer",
            "title": "Year"
          },
          "trim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trim"
          },
          "vin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vin"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          },
          "user_id": {
            "type": "integer",
            "title": "User Id"
          }
        },
        "type": "object",
        "required": [
          "id",
          "make",
          "model",
          "year",
          "user_id"
        ],
        "title": "CarRead"
      },
      "CarUpdate": {
        "properties": {
          "make": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Make"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          },
          "trim": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trim"
          },
          "vin": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vin"
          },
          "image_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Url"
          }
        },
        "type": "object",
        "title": "CarUpdate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}